requirements.txt
telethon
supabase
tweepy
google-generativeai
feedparser
python-dotenv
requests
fuzzywuzzy
python-Levenshtein
